Refactored the importing functions
Separated out and renamed the existing classes for our importing functions.
This is meant to standardize on the interface names to make it easier to add
new package and advisory types.
