+++ app/models/import/packages.rb
@@ -1,0 +1,3 @@
+# Empty class to be inherited later.
+class Import::Packages
+end
