+++ lib/tasks/import/rhel_advisories.rake
@@ -1,6 +1,6 @@
 namespace :import do
   desc 'Import rhel advisories from the web'
   task rhel_advisories: :environment do
-    Import::Yum.new.rhel_advisories
+    Import::Packages::Yum::RHEL.new.import_advisories
   end
 end
