+++ lib/tasks/import/ruby_advisory_db.rake
@@ -1,7 +1,7 @@
 namespace :import do
   desc 'Import gem advisories from the ruby advisory database'
   task ruby_advisory_db: :environment do
-    Import::Gems.new.update_source
-    Import::Gems.new.ruby_advisories
+    Import::Packages::Gems.new.update_source
+    Import::Packages::Gems.new.import_advisories
   end
 end
